requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
google-auth==2.23.4
google-cloud-dialogflow-cx==1.28.0
protobuf==4.25.1
